chore: skip PR on publish (#26)
We were also accidentally creating an OUT= file on the publish sequence which would then cause a PR to be opened. We have added a new output that we can use to skip PR straight away unless it is the version command. This fixes the problem, but avoids the root issue of the OUT= file for now.