Add Tauri 2.0 Action Script
Tauri 2.0 requires webkitgtk 4.1 and ubuntu 22.04 should now be used. NPM projects have the CLI as part of the project where it needs to be installed with Cargo. We manually specified the Tauri version to be able to get Tauri 2.0. If you are signing the Tauri Release, you need to specify the keys and passwords in the Secrets setting in GitHub.