Allow building binaries without artifacts
Before this commit, tauri-action refused to create such binaries.

After the commit, binaries without artifacts are permitted as long as
uploading of the artifacts was not requested. I'm gathering that this is
the way to do it from the last paragraph of the README:

> If you only want to build the app without having the action upload any
> assets, for example if you want to only use actions/upload-artifact,
> simply omit tagName, releaseName and releaseId.

I wanted to produce binaries, but not bundles, and disabled bundles in
my tauri.conf.json:

https://github.com/ilyagr/diffedit3/blob/cce71325442edc1d938b32261878ea50e5401a49/backend-tauri/tauri.conf.json#L37-L40

I'm currently using this patch here:

https://github.com/ilyagr/diffedit3/blob/cce71325442edc1d938b32261878ea50e5401a49/.github/workflows/release.yml#L112-L114